imitial commit

+++ src/main/java/ru/netology/stats/StatService.java
@@ -40,7 +40,7 @@ public class StatService {
     public int lowSalesMonths(int[] sales) {
         int lowSalesMonths = 0;
         for (int i = 0; i < sales.length; i++) {
-            if (sales[(int) i] < allSalesAverageSum(sales)) {
+            if (sales[i] < allSalesAverageSum(sales)) {
                 lowSalesMonths++;
             }
         }
@@ -50,7 +50,7 @@ public class StatService {
     public int highSalesMonths(int[] sales) {
         int highSalesMonths = 0;
         for (int i = 0; i < sales.length; i++) {
-            if (sales[(int) i] > allSalesAverageSum(sales)) {
+            if (sales[i] > allSalesAverageSum(sales)) {
                 highSalesMonths++;
             }
         }
